Add error module to common
This error type is ment to directly map to HTTP status codes and be used
by the resources contained in `resources`.

Any actix version that relies on resources using these error types is
going to need to implement their own error type as an intermediary.
Actix-web has traits for converting errors into responses but because of
rusts foreign implementation requirements each crate can not
implement those traits for this type.

Signed-off-by: Caleb Hill <hill@bitwise.io>