Add protocol version to actix_web_4 crate
Protocol version arguably could belong in common because the
protocol_versions should be the same between actix crates. My reasoning
for splitting it out is twofold.

1. The enum would have to be duplicated in each actix-crate anyways if
we wanted to use any of the actix traits which is not useful
duplication, just technical.

2. Making it a crate implementation detail makes building crates with
only a subset of the protocol versions easier withouth breaking from the
general implementation pattern.

As an example, if we wanted to build a lighterweight network with only
the newest versions of everything the protocol_version implementation
could be as simple as
`
// Please note this does not compile and is just an example.
struct ProtocolVersion{}
impl TryFrom<Request> for ProtocolVersion {
    type Error = RestError;
    fn try_from(value: Request) -> Result<Self,Self::Error> {
        match value.get("protocol_version") {
            Some(val) => match val{
                3 => Ok(Self{}),
                _ =>  Err(RestError::UnsupportedProtocolVersion)
            },
            None => Err(RestError::BadRequest)
        }
}
`

Instead of having to write an intermediary that both converts between a
`Request` struct, a `common::ProtocolVersion`, and then breaks the
general established pattern of mirroring the `common::ProtocolVersion`
enum in the child crate.

Signed-off-by: Caleb Hill <hill@bitwise.io>