Add get_admin_circuits V1Response struct
When the v1 get_admin_circuits code is called from the common crate it
returns a distinct struct. The ideal case is that we then directly
convert that to a http compatible representation and send it down the
wire. Actix-web has a trait to convert between an internal struct and an
HttpResponse, however since the type we want to convert to an
HttpResponse is not from the splinter_rest_api_actix_web_4 crate we
can't directly implement that trait.

This commit adds an intermediary struct that wraps the response type and
implements `Responder`.