Add wrapper Method type to use with PermissionMap
Since PermissionMap takes a generic type that can convert to its
internal Method type, and you have to convert from the actix-web Method
there needs to be an intermediary. Rust does not allow implementing
traits on and for foreign types so this commit introduces a wrapper
struct with a single field that we can then implement `Into` on.

While I don't love this solution its kinda unavoidable with the crate
boundaries being what they are.

The alternative is to make PermissionMap and RequestDefinition generic
over Method types. Which is maybe a better long term plan.