Add builder module to actix_web_4
The builder exposes a much smaller set of methods than previous
versions. In previous implementations each resource provider was
provided in a seperate usually feature gated step. That design made more
sense when everything was in one crate, reimplementing that same pattern
now would involve importing every resource provider, and the builder and
then just providing them to the builder.

This builder moves the resource provider parts of that to this crate
where they are mostly configured by features in the `build` method. The
methods that are public provide store factories and other instance
dependent objects.

There is an argument to be made for adding a `RestApiBuilder` trait in
common because most of these methods will be common between the actix-1,
actix-4, and any future implementations. The tricky thing then becomes
what to return from `build` and `run`. For the moment I won't be adding
a trait to capture that behaviour.