Update ProposalStore to follow clone_boxed pattern
Having a trait require `Clone` introduces a `Sized` trait requirement
which makes it difficult wrap the trait object in any sort of pointer
type. This commit updates the `ProposalStore` trait by removing the
`Clone` trait bound, including a `clone_boxed` method and updating all
current implementations and usages of the trait to use a boxed trait
instead of generics.

Signed-off-by: Caleb Hill <hill@bitwise.io>